moved vessel details in own component
